Move presenter screenshots to their own modules
Announcements and Dictionary presenter screenshot tests are moved from `PresenterFullScreenScreenshotTest` in `:composeApp` into dedicated `AnnouncementsPresenterScreenshotTest` and `DictionaryPresenterScreenshotTest` classes in `:announcements-tab` and `:dictionary-tab` respectively. The golden images move with them. CI workflow updated to record and upload both new module screenshot directories.
