+++ dictionary-tab/screenshots/dictionaryPresenter/dictionary_entry.png
