+++ dictionary-tab/screenshots/dictionaryPresenter/dictionary_entry_styled.png
